+++ Source/Cloud9/Weapon/Classes/Cloud9WeaponGrenade.h
@@ -45,7 +45,7 @@ public:
 
 protected:
 	virtual bool OnInitialize(const FWeaponId& NewWeaponId, FName NewWeaponSkin) override;
-	virtual bool DeInitialize() override;
+	virtual void DeInitialize() override;
 
 protected:
 	/**
