+++ Source/Cloud9/Weapon/Classes/Cloud9WeaponMelee.h
@@ -51,7 +51,7 @@ public:
 
 protected:
 	virtual bool OnInitialize(const FWeaponId& NewWeaponId, FName NewWeaponSkin) override;
-	virtual bool DeInitialize() override;
+	virtual void DeInitialize() override;
 	virtual void Tick(float DeltaSeconds) override;
 
 protected:
