+++ Source/Cloud9/Weapon/Enums/WeaponId.h
@@ -28,75 +28,5 @@
 #include "GrenadeNames.h"
 #include "MeleeNames.h"
 #include "UnknownNames.h"
-#include "Cloud9/Tools/Extensions/UEnum.h"
-#include "Cloud9/Tools/Macro/Operator.h"
-
-class ACloud9WeaponGrenade;
-class ACloud9WeaponMelee;
-class ACloud9WeaponFirearm;
-
-struct FGrenadeWeaponInfo;
-struct FMeleeWeaponInfo;
-struct FFirearmWeaponInfo;
 
 using FWeaponId = TVariant<EBadWeapon, EMelee, EFirearm, EGrenade>;
-
-namespace EFWeaponId
-{
-	struct ToName
-	{
-	public:
-		FORCEINLINE FName operator()(const FWeaponId& Self) const
-		{
-			return Visit([this](let It) { return It | EUEnum::GetValueName(); }, Self);
-		}
-
-		OPERATOR_BODY(ToName)
-	};
-
-	/* WeaponInfo */
-
-	template <typename WeaponIdType>
-	struct WeaponInfo;
-
-	template <>
-	struct WeaponInfo<EFirearm>
-	{
-		using Type = FFirearmWeaponInfo;
-	};
-
-	template <>
-	struct WeaponInfo<EMelee>
-	{
-		using Type = FMeleeWeaponInfo;
-	};
-
-	template <>
-	struct WeaponInfo<EGrenade>
-	{
-		using Type = FGrenadeWeaponInfo;
-	};
-
-	/* WeaponClass */
-
-	template <typename WeaponIdType>
-	struct WeaponClass;
-
-	template <>
-	struct WeaponClass<EFirearm>
-	{
-		using Type = ACloud9WeaponFirearm;
-	};
-
-	template <>
-	struct WeaponClass<EMelee>
-	{
-		using Type = ACloud9WeaponMelee;
-	};
-
-	template <>
-	struct WeaponClass<EGrenade>
-	{
-		using Type = ACloud9WeaponGrenade;
-	};
-}
